Rename scripts folder to src, create index.js file
