+++ src/chart-functionality.js
@@ -1,4 +1,4 @@
-"use strict";
+import { horiz1Bars } from './index.js';
 
 // Drop Down Menu Logic
 const charts = document.getElementsByClassName("charts");
@@ -60,15 +60,6 @@ selectElement.addEventListener("mouseleave", () => {
         lightSelect.style.opacity = "0";
     }
 });
-
-// Chart Variables
-const botSpace = `<div height="24" style="font-size: 24px; line-height: 24px;">&nbsp;</div>`; // Bottom spacing for non-vertical charts
-const shadowColor = "0 0 10px #ffdf65, 0 0 10px #ffdf65, 0 0 10px #ffffff, 0 0 10px #ffffff"; // Chart text highlighting
-const barSizeShadow = "-1px -1px 10px #ffdf65, -1px 1px 10px #ffdf65"; // Data bar size highlighting
-const barColorShadow = "-1px -1px 10px #6585ff, -1px 1px 10px #6585ff, 0 0 10px #ffffff"; // Data bar color highlighting
-const imgDropShadow = "drop-shadow(0 0 8px #ffdf65) drop-shadow(0 0 8px #ffdf65) drop-shadow(0 0 8px #ffffff)"; // Image drop-shadow
-const msoOpen = `<!--[if mso 15 | mso 16]><table align="center" border="0" cellpadding="0" cellspacing="0" width="600"><tr><td><![endif]-->`;
-const msoClose = "<!--[if mso 15 | mso 16]></td></tr></table><![endif]-->";
 
 // Preview Dark Mode
 const lightMode = document.getElementById("lightMode");
